Moved topology helpers to where they're used
